Check whether the tile dir exists before trying to delete it.
git-svn-id: https://svn.code.sf.net/p/vassalengine/svn/VASSAL-src/trunk@8426 67b53d14-2c14-4ace-a08f-0dab2b34000c

+++ src/VASSAL/launch/ModuleManagerWindow.java
@@ -1359,11 +1359,13 @@ public class ModuleManagerWindow extends JFrame {
       );
 
       final File tdir = new File(Info.getConfDir(), "tiles/" + hstr);
-      try {
-        FileUtils.forceDelete(tdir);
-      }
-      catch (IOException e) {
-        WriteErrorDialog.error(e, tdir);
+      if (tdir.exists()) {
+        try {
+          FileUtils.forceDelete(tdir);
+        }
+        catch (IOException e) {
+          WriteErrorDialog.error(e, tdir);
+        }
       }
     }
